refactor(region): extract RepresentativesSyncService — #828 Step 3
Third bounded-context split. Extracts syncRepresentatives + its two
helpers (sanitizeDistrict, normalizeRep) out of RegionSyncService.

region-sync.service.ts shrinks from 3,318 → 3,120 LOC (cumulative
3,929 → 3,120, −21% so far). RepresentativesSyncService is 299 LOC.

The extracted service takes its full dep graph via constructor injection
(DbService + @Optional bioGenerator + committeeSummaryGenerator +
legislativeCommitteeLinker + legislativeCommitteeDescriptions). The
orchestrator passes plugin-scoped context (regionName +
normalizeDistrict + bioNoisePatterns) per-call via a RepresentativesPluginContext
object rather than reaching back through DI — the plugin state stays on
the orchestrator until #828 Step 6 (plugin lifecycle extraction).

`deriveDistrictFromExternalId` was defined in BOTH region.service.ts and
region-sync.service.ts; the duplicate is removed here and the
orchestrator imports the canonical one from region.service.ts to pass
into the extracted service. Behavior unchanged.

All 2,105 backend tests pass.

Refs #828
